queue reducer refactor
