+++ app/main.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 import {Router, Route, IndexRedirect, browserHistory} from 'react-router'
 import {render} from 'react-dom'
 import {connect, Provider} from 'react-redux'
-import {fetchQueue} from 'APP/app/utils/queue'
+import {fetchQueue} from 'APP/app/reducers/queue'
 import {fetchSetItems} from 'APP/app/reducers/set'
 import store from 'APP/app/store'
 import {Output} from 'APP/app/components/Output'
