+++ app/reducers/index.jsx
@@ -6,8 +6,7 @@ const rootReducer = combineReducers({
   playerLeft: require ('./playerLeft').default,
   playerRight: require ('./playerRight').default,
   // queue: require ('./queue').default
-  queueLeft: require('./queueLeft').default,
-  queueRight: require('./queueRight').default,
+  queue: require('./queue').default,
 //player: require('./player'), // for both players
 //queue: require('./queue'), // for both queues
   set: require('./set').default,
