+++ app/reducers/queue.jsx
@@ -1,27 +1,61 @@
-// we have to rename everything else for this to work so leaving this commented out rn
+import localStore from 'store'
 
-// import {combineReducers} from 'redux'
-// import {RECEIVE_LEFT, RECEIVE_RIGHT} from '../utils/queue'
+// CONSTANTS
+const RECEIVE_LEFT = 'RECEIVE_LEFT'
+const RECEIVE_RIGHT = 'RECEIVE_RIGHT'
 
-// const queue = (state = [], action) => {
-//  switch (action.type) {
-//    case RECEIVE_LEFT:
-//      return action.queueLeft
-//  }
-//  return state
-// }
+//REDUCER
+const reducer = (state = {Left: [], Right: []}, action) => {
+  switch (action.type) {
+    case RECEIVE_LEFT:
+      return Object.assign({}, state, {Left: action.queueLeft})
+    case RECEIVE_RIGHT:
+      return Object.assign({}, state, {Right: action.queueRight})
+    default:
+      return state
+  }
+}
 
-// const queueRight = (state = [], action) => {
-//  switch (action.type) {
-//    case RECEIVE_RIGHT:
-//      return action.queueRight
-//  }
-//  return state
-// }
+// ACTION TYPES
+export const receiveQueue = (queue, queueLeftOrRight) => ({
+  type: queueLeftOrRight === 'queueLeft' ? RECEIVE_LEFT : RECEIVE_RIGHT,
+  [queueLeftOrRight]: queue
+})
 
-// const queueReducer = combineReducers({
-//  left: queueLeft,
-//  right: queueRight
-// })
+// ACTION CREATORS
+export const fetchQueue = queueLeftOrRight => dispatch => {
+  dispatch(receiveQueue(localStore.get(queueLeftOrRight) || [], queueLeftOrRight))
+  // console.log(`${queueLeftOrRight}:`, getState()[queueLeftOrRight])
+}
 
-// export default queueReducer
+export const addToQueue = (video, queueLeftOrRight) => dispatch => {
+  // get the correct queue
+  const queueToUpdate = localStore.get(queueLeftOrRight) || []
+  // update queue
+  queueToUpdate.push(video)
+  // set new queue
+  localStore.set(queueLeftOrRight, queueToUpdate)
+  // put new queue on state
+  dispatch(receiveQueue(queueToUpdate, queueLeftOrRight))
+}
+
+export const removeFromQueue = (videoIdx, queueLeftOrRight) => dispatch => {
+  // get the correct queue
+  const queueToUpdate = localStore.get(queueLeftOrRight) || []
+  // update queue
+  queueToUpdate.splice(videoIdx, 1)
+  // set new queue
+  localStore.set(queueLeftOrRight, queueToUpdate)
+  // put new queue on state
+  dispatch(fetchQueue(queueLeftOrRight))
+}
+
+export const clearQueue = queueLeftOrRight => dispatch => {
+  localStore.set(queueLeftOrRight, [])
+  dispatch(receiveQueue([], queueLeftOrRight))
+}
+
+export default reducer
+// move stuff around: combination of remove from queue and add to queue
+
+// play from back in queue: combination of remove and add
